requirements.txt
python-dotenv
flask
flask_socketio
python-dateutil
openai
requests
google-generativeai
opencc-python-reimplemented